Reformat Code,修改ready文档，修改文件夹页面的路径文案样式

+++ vitamio/src/main/java/io/vov/vitamio/Vitamio.java
@@ -17,12 +17,10 @@
 package io.vov.vitamio;
 
 import android.content.Context;
-
 import io.vov.vitamio.utils.CPU;
 import io.vov.vitamio.utils.ContextUtils;
 import io.vov.vitamio.utils.IOUtils;
 import io.vov.vitamio.utils.Log;
-
 import java.io.BufferedInputStream;
 import java.io.BufferedReader;
 import java.io.File;
@@ -41,12 +39,13 @@ import java.util.List;
  * Don't modify this class, or the full Vitamio library may be broken.
  */
 public class Vitamio {
-  private static final String[] LIBS_ARM_CODECS = {"libvvo.7.so", "libvvo.8.so", "libffmpeg.so", "libOMX.9.so", "libOMX.11.so", "libOMX.14.so", "libOMX.18.so"};
-  private static final String[] LIBS_X86_CODECS = {"libffmpeg.so", "libOMX.9.so", "libOMX.14.so", "libOMX.18.so"};
-  private static final String[] LIBS_MIPS_CODECS = {"libffmpeg.so", "libOMX.14.so"};
-  private static final String[] LIBS_PLAYER = {"libvplayer.so"};
-  private static final String[] LIBS_SCANNER = {"libvscanner.so"};
-  private static final String[] LIBS_AV = {"libvao.0.so", "libvvo.0.so", "libvvo.9.so", "libvvo.j.so"};
+  private static final String[] LIBS_ARM_CODECS =
+      { "libvvo.7.so", "libvvo.8.so", "libffmpeg.so", "libOMX.9.so", "libOMX.11.so", "libOMX.14.so", "libOMX.18.so" };
+  private static final String[] LIBS_X86_CODECS = { "libffmpeg.so", "libOMX.9.so", "libOMX.14.so", "libOMX.18.so" };
+  private static final String[] LIBS_MIPS_CODECS = { "libffmpeg.so", "libOMX.14.so" };
+  private static final String[] LIBS_PLAYER = { "libvplayer.so" };
+  private static final String[] LIBS_SCANNER = { "libvscanner.so" };
+  private static final String[] LIBS_AV = { "libvao.0.so", "libvvo.0.so", "libvvo.9.so", "libvvo.j.so" };
   private static final String LIBS_LOCK = ".lock";
   private static final int VITAMIO_NOT_SUPPORTED = -1;
   private static final int VITAMIO_MIPS = 40;
@@ -59,20 +58,21 @@ public class Vitamio {
 
   static {
     int cpu = CPU.getFeature();
-    if ((cpu & CPU.FEATURE_ARM_NEON) > 0)
+    if ((cpu & CPU.FEATURE_ARM_NEON) > 0) {
       vitamioType = VITAMIO_ARMV7_NEON;
-    else if ((cpu & CPU.FEATURE_ARM_VFPV3) > 0 && (cpu & CPU.FEATURE_ARM_V7A) > 0)
+    } else if ((cpu & CPU.FEATURE_ARM_VFPV3) > 0 && (cpu & CPU.FEATURE_ARM_V7A) > 0) {
       vitamioType = VITAMIO_ARMV7_VFPV3;
-    else if ((cpu & CPU.FEATURE_ARM_VFP) > 0 && (cpu & CPU.FEATURE_ARM_V6) > 0)
+    } else if ((cpu & CPU.FEATURE_ARM_VFP) > 0 && (cpu & CPU.FEATURE_ARM_V6) > 0) {
       vitamioType = VITAMIO_ARMV6_VFP;
-    else if ((cpu & CPU.FEATURE_ARM_V6) > 0)
-    	vitamioType = VITAMIO_ARMV6;
-    else if ((cpu & CPU.FEATURE_X86) > 0)
-    	vitamioType = VITAMIO_X86;
-    else if ((cpu & CPU.FEATURE_MIPS) > 0) 
-    	vitamioType = VITAMIO_MIPS;
-    else
-    	vitamioType = VITAMIO_NOT_SUPPORTED;
+    } else if ((cpu & CPU.FEATURE_ARM_V6) > 0) {
+      vitamioType = VITAMIO_ARMV6;
+    } else if ((cpu & CPU.FEATURE_X86) > 0) {
+      vitamioType = VITAMIO_X86;
+    } else if ((cpu & CPU.FEATURE_MIPS) > 0) {
+      vitamioType = VITAMIO_MIPS;
+    } else {
+      vitamioType = VITAMIO_NOT_SUPPORTED;
+    }
   }
 
   private static String vitamioPackage;
@@ -94,7 +94,7 @@ public class Vitamio {
   /**
    * Same as {@link #initialize(Context)}
    *
-   * @param ctx   Android Context
+   * @param ctx Android Context
    * @param rawId R.raw.libarm
    * @return true if the Vitamio initialized successfully.
    */
@@ -123,18 +123,17 @@ public class Vitamio {
           }
         }
         File lock = new File(getLibraryPath() + LIBS_LOCK);
-        BufferedReader buffer = null; 
+        BufferedReader buffer = null;
         try {
-          buffer = new BufferedReader(new FileReader(lock));  
+          buffer = new BufferedReader(new FileReader(lock));
           int appVersion = ContextUtils.getVersionCode(ctx);
-          int libVersion = Integer.valueOf(buffer.readLine());  
+          int libVersion = Integer.valueOf(buffer.readLine());
           Log.i("isNativeLibsInited, APP VERSION: %d, Vitamio Library version: %d", appVersion, libVersion);
-          if (libVersion == appVersion)
-            return true;
+          if (libVersion == appVersion) return true;
         } catch (IOException e) {
           Log.e("isNativeLibsInited", e);
         } catch (NumberFormatException e) {
-        	Log.e("isNativeLibsInited", e);
+          Log.e("isNativeLibsInited", e);
         } finally {
           IOUtils.closeSilently(buffer);
         }
@@ -159,23 +158,22 @@ public class Vitamio {
     List<String> libs = new ArrayList<String>();
     String[][] vitamioLibs = null;
     switch (vitamioType) {
-		case VITAMIO_ARMV6:
-		case VITAMIO_ARMV6_VFP:
-		case VITAMIO_ARMV7_VFPV3:
-		case VITAMIO_ARMV7_NEON:
-			vitamioLibs = new String[][]{LIBS_ARM_CODECS, LIBS_PLAYER, LIBS_SCANNER, LIBS_AV};
-			break;
-		case VITAMIO_X86:
-			vitamioLibs = new String[][]{LIBS_X86_CODECS, LIBS_PLAYER, LIBS_SCANNER, LIBS_AV};
-			break;
-		case VITAMIO_MIPS:
-			vitamioLibs = new String[][]{LIBS_MIPS_CODECS, LIBS_PLAYER, LIBS_SCANNER, LIBS_AV};
-			break;
-		default:
-			break;
-		}
-    if (vitamioLibs == null)
-    	return libs;
+      case VITAMIO_ARMV6:
+      case VITAMIO_ARMV6_VFP:
+      case VITAMIO_ARMV7_VFPV3:
+      case VITAMIO_ARMV7_NEON:
+        vitamioLibs = new String[][] { LIBS_ARM_CODECS, LIBS_PLAYER, LIBS_SCANNER, LIBS_AV };
+        break;
+      case VITAMIO_X86:
+        vitamioLibs = new String[][] { LIBS_X86_CODECS, LIBS_PLAYER, LIBS_SCANNER, LIBS_AV };
+        break;
+      case VITAMIO_MIPS:
+        vitamioLibs = new String[][] { LIBS_MIPS_CODECS, LIBS_PLAYER, LIBS_SCANNER, LIBS_AV };
+        break;
+      default:
+        break;
+    }
+    if (vitamioLibs == null) return libs;
     for (String[] libArray : vitamioLibs) {
       for (String lib : libArray)
         libs.add(lib);
@@ -189,8 +187,7 @@ public class Vitamio {
     final int version = ContextUtils.getVersionCode(ctx);
     Log.d("loadLibs start " + version);
     File lock = new File(getLibraryPath() + LIBS_LOCK);
-    if (lock.exists())
-      lock.delete();
+    if (lock.exists()) lock.delete();
     String libPath = copyCompressedLib(ctx, rawID, "libarm.so");
     Log.d("copyCompressedLib time: " + (System.currentTimeMillis() - begin) / 1000.0);
     boolean inited = native_initializeLibs(libPath, getLibraryPath(), String.valueOf(Vitamio.getVitamioType()));
@@ -223,15 +220,11 @@ public class Vitamio {
         String destDir = getLibraryPath();
         destPath = destDir + destName;
         File f = new File(destDir);
-        if (f.exists() && !f.isDirectory())
-          f.delete();
-        if (!f.exists())
-          f.mkdirs();
+        if (f.exists() && !f.isDirectory()) f.delete();
+        if (!f.exists()) f.mkdirs();
         f = new File(destPath);
-        if (f.exists() && !f.isFile())
-          f.delete();
-        if (!f.exists())
-          f.createNewFile();
+        if (f.exists() && !f.isFile()) f.delete();
+        if (!f.exists()) f.createNewFile();
       } catch (Exception fe) {
         Log.e("loadLib", fe);
       }
